requirements.txt
python-dotenv==1.2.1
livekit-plugins-openai==1.2.18
livekit-agents>=1.2.18
openai>=2.7.1
aiohttp~=3.10
av>=14.0.0,<16
